Fix OWIN middleware order
This commit will place the line

    app.UseWebApi(config);

last in the Startup Configuration method.

The reason for this is that the order that you declare your OWIN
middleware instances is important.  UseWebApi terminates the
pipeline and should always be last, regardless.

But also, app.UseJwtBearerAuthentication should always preceed
any other middlewares that need the user to be authenticated, like
logging, /or/ Web API itself.